test: migrate dev error tests into separate sandbox test
